LocalEats/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.1.0.post1
Werkzeug==3.0.1
python-dotenv==1.0.1
bcrypt==4.0.1
requests==2.31.0
